tests: Condition np_notif_store_test with the in-build-dir repo location
Fixes: https://travis-ci.org/sysrepo/sysrepo/jobs/391867403
Cc: https://github.com/sysrepo/sysrepo/issues/1148
