+++ tests/np_test.c
@@ -404,55 +404,6 @@ np_dp_subscriptions_test(void **state)
     assert_int_equal(rc, SR_ERR_OK);
 }
 
-static void
-np_notif_store_test(void **state)
-{
-#ifndef ENABLE_NOTIF_STORE
-    skip();
-#else
-    int rc = SR_ERR_OK;
-    test_ctx_t *test_ctx = *state;
-    assert_non_null(test_ctx);
-    np_ctx_t *np_ctx = test_ctx->rp_ctx->np_ctx;
-
-    struct ly_ctx *ctx = NULL;
-    const struct lys_module *module = NULL;
-    struct lyd_node *node = NULL;
-    sr_list_t *notif_list = NULL;
-
-    /* create notif. data tree */
-    ctx = ly_ctx_new(TEST_SCHEMA_SEARCH_DIR, 0);
-    assert_non_null(ctx);
-    module = ly_ctx_load_module(ctx, "test-module", NULL);
-    assert_non_null(module);
-    node = lyd_new_path(NULL, ctx, "/test-module:link-discovered/source/interface", "eth0", 0, 0);
-    assert_non_null(node);
-
-    /* store notification */
-    rc = np_store_event_notification(np_ctx, test_ctx->rp_session_ctx->user_credentials, "/test-module:link-discovered",
-            time(NULL), node);
-    assert_int_equal(rc, SR_ERR_OK);
-
-    /* retrieve notifications  */
-    rc = np_get_event_notifications(np_ctx, test_ctx->rp_session_ctx, "/test-module:link-discovered", 0, time(NULL),
-            SR_API_VALUES, &notif_list);
-    assert_int_equal(rc, SR_ERR_OK);
-    assert_non_null(notif_list);
-
-    for (size_t i = 0; i < notif_list->count; i++) {
-        np_ev_notification_t *notification = notif_list->data[i];
-        assert_string_equal(notification->xpath, "/test-module:link-discovered");
-        np_event_notification_cleanup(notification);
-    }
-    sr_list_cleanup(notif_list);
-
-    rc = np_notification_store_cleanup(np_ctx, false);
-    assert_int_equal(rc, SR_ERR_OK);
-    lyd_free_withsiblings(node);
-    ly_ctx_destroy(ctx, NULL);
-#endif
-}
-
 int
 main() {
     const struct CMUnitTest tests[] = {
@@ -462,7 +413,6 @@ main() {
             cmocka_unit_test_setup_teardown(np_hello_notify_test, test_setup, test_teardown),
             cmocka_unit_test_setup_teardown(np_module_subscriptions_test, test_setup, test_teardown),
             cmocka_unit_test_setup_teardown(np_dp_subscriptions_test, test_setup, test_teardown),
-            cmocka_unit_test_setup_teardown(np_notif_store_test, test_setup, test_teardown),
     };
 
     watchdog_start(300);
